# front/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "E-commerce Platform API",
    "version": "1.0.0",
    "description": "Complete e-commerce platform with comprehensive product management, order processing, payment integration, and user management",
    "contact": {
      "name": "API Support",
      "email": "support@ecommerce-platform.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://api.ecommerce-platform.com",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token for authentication. Get token from login endpoint and use format: Bearer <token>",
        "example": "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
      },
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key for service-to-service authentication"
      }
    },
...
    "schemas": {
...
      "WishlistItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "userId": {
            "type": "integer"
          },
          "productId": {
            "type": "integer"
          },
          "product": {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer"
              },
              "name": {
                "type": "string"
              },
              "slug": {
                "type": "string"
              },
              "price": {
                "type": "integer"
              },
              "discountPrice": {
                "type": "integer"
              },
              "effectivePrice": {
                "type": "integer"
              },
              "stock": {
                "type": "integer"
              },
              "availableStock": {
                "type": "integer"
              },
              "averageRating": {
                "type": "number"
              },
              "reviewCount": {
                "type": "integer"
              },
              "mainImage": {
                "type": "string"
              }
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    },
...
  },
...
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/wishlist": {
      "post": {
        "summary": "Add product to wishlist",
        "tags": [
          "Wishlist"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "productId"
                ],
                "properties": {
                  "productId": {
                    "type": "integer",
                    "description": "Product ID to add to wishlist"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Product added to wishlist successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/WishlistItem"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error or product already in wishlist"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Product not found"
          }
        }
      },
      "get": {
        "summary": "Get user's wishlist",
        "tags": [
          "Wishlist"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "description": "Page number"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50,
              "default": 20
            },
            "description": "Items per page"
          },
          {
            "in": "query",
            "name": "sortBy",
            "schema": {
              "type": "string",
              "enum": [
                "createdAt",
                "productName",
                "price"
              ],
              "default": "createdAt"
            },
            "description": "Sort field"
          },
          {
            "in": "query",
            "name": "sortOrder",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ],
              "default": "desc"
            },
            "description": "Sort order"
          }
        ],
        "responses": {
          "200": {
            "description": "Wishlist retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "items": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/WishlistItem"
                          }
                        },
                        "pagination": {
                          "type": "object"
                        },
                        "stats": {
                          "type": "object",
                          "properties": {
                            "totalItems": {
                              "type": "integer"
                            },
                            "availableItems": {
                              "type": "integer"
                            },
                            "outOfStockItems": {
                              "type": "integer"
                            },
                            "estimatedValue": {
                              "type": "integer"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/v1/wishlist/{id}": {
      "delete": {
        "summary": "Remove product from wishlist",
        "tags": [
          "Wishlist"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Wishlist item ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Product removed from wishlist successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Wishlist item not found"
          }
        }
      }
    },
    "/api/v1/wishlist/{id}/move-to-cart": {
      "post": {
        "summary": "Move wishlist item to cart",
        "tags": [
          "Wishlist"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Wishlist item ID"
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "quantity": {
                    "type": "integer",
                    "minimum": 1,
                    "default": 1,
                    "description": "Quantity to add to cart"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Product moved to cart successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "cartItem": {
                          "type": "object"
                        },
                        "message": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error or insufficient stock"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Wishlist item not found"
          }
        }
      }
    },
    "/api/v1/wishlist/clear": {
      "delete": {
        "summary": "Clear entire wishlist",
        "tags": [
          "Wishlist"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Wishlist cleared successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "removedCount": {
                          "type": "integer"
                        },
                        "message": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  }
}